requirements.txt
aiohttp==3.9.3
aiosignal==1.3.1
aniso8601==9.0.1
annotated-types==0.6.0
anyio==4.2.0
asttokens==2.4.1
attrs==23.2.0
beautifulsoup4==4.12.3
blinker==1.7.0
bs4==0.0.2
cachetools==5.3.2
certifi==2023.11.17
charset-normalizer==3.3.2
ci-info==0.3.0
click==8.1.7
colorama==0.4.6
comm==0.2.1
configobj==5.0.8
configparser==6.0.0
dataclasses-json==0.6.4
debugpy==1.6.7
decorator==5.1.1
distro==1.9.0
etelemetry==0.3.1
exceptiongroup==1.2.0
executing==2.0.1
filelock==3.13.1
Flask==3.0.1
flask-restx==1.3.0
frozenlist==1.4.1
google-api-core==2.16.2
google-api-python-client==2.116.0
google-auth==2.27.0
google-auth-httplib2==0.2.0
googleapis-common-protos==1.62.0
greenlet==3.0.3
h11==0.14.0
httpcore==1.0.2
httplib2==0.22.0
httpx==0.26.0
idna==3.6
importlib-metadata==7.0.1
importlib-resources==6.1.1
ipykernel==6.29.0
ipython==8.20.0
isodate==0.6.1
itsdangerous==2.1.2
jedi==0.19.1
Jinja2==3.1.3
jsonpatch==1.33
jsonpointer==2.4
jsonschema==4.21.1
jsonschema-specifications==2023.12.1
jupyter_client==8.6.0
jupyter_core==5.5.0
langchain==0.1.5
langchain-community==0.0.17
langchain-core==0.1.18
langsmith==0.0.86
looseversion==1.3.0
lxml==5.1.0
MarkupSafe==2.1.4
marshmallow==3.20.2
matplotlib-inline==0.1.6
multidict==6.0.5
mypy-extensions==1.0.0
nest_asyncio==1.6.0
networkx==3.2.1
nibabel==5.2.0
nipype==1.8.6
numpy==1.26.3
openai==1.10.0
packaging==23.2
pandas==2.2.0
parso==0.8.3
pathlib==1.0.1
pickleshare==0.7.5
pinecone-client==3.0.2
pip==23.3.1
platformdirs==4.1.0
prompt-toolkit==3.0.42
protobuf==4.25.2
prov==2.0.0
psutil==5.9.0
pure-eval==0.2.2
pyasn1==0.5.1
pyasn1-modules==0.3.0
pydantic==2.6.0
pydantic_core==2.16.1
pydot==2.0.0
Pygments==2.17.2
PyMuPDF==1.23.21
PyMuPDFb==1.23.9
pyparsing==3.1.1
python-dateutil==2.8.2
python-dotenv==1.0.1
python-magic-bin==0.4.14
pytz==2023.4
pywin32==305.1
pyxnat==1.6.2
PyYAML==6.0.1
pyzmq==25.1.2
rdflib==7.0.0
referencing==0.33.0
requests==2.31.0
rpds-py==0.17.1
rsa==4.9
scipy==1.12.0
setuptools==68.2.2
simplejson==3.19.2
six==1.16.0
sniffio==1.3.0
soupsieve==2.5
SQLAlchemy==2.0.25
stack-data==0.6.2
tenacity==8.2.3
tornado==6.3.3
tqdm==4.66.1
traitlets==5.14.1
traits==6.3.2
typing_extensions==4.9.0
typing-inspect==0.9.0
tzdata==2023.4
uritemplate==4.1.1
urllib3==2.2.0
wcwidth==0.2.13
Werkzeug==3.0.1
wheel==0.41.2
yarl==1.9.4
zipp==3.17.0